New update instance script
We are moving most of the update functionality to the
infrastructure-live repo. This new update script will be called from
that repo to update the given instance. It contains the parts of the
update procedure which are dependent of the internal folder structure
of vm-images repo. This is why it is kept here and not moved to
infrastructure-live -- to preserve separation of concerns.